requirements.txt
langgraph==1.0.5
langgraph-prebuilt==1.0.5
python-dotenv
langsmith
langchain-openai
langchain-community
langchain-tavily
langchain-experimental
tavily-python 
pandas 
openai
requests
azure-identity
IPython
azure-ai-agentserver-agentframework
azure-ai-agentserver-langgraph==1.0.0b7
azure-monitor-opentelemetry==1.8.1
